backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
pymongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
boto3==1.28.44
openai==0.28.1
gTTS==2.3.2
pydub==0.25.1
SpeechRecognition==3.10.0
PyAudio==0.2.11
pandas==2.1.3
openpyxl==3.1.2
pytest==7.4.3
gunicorn==21.2.0
Werkzeug==2.3.7
python-multipart==0.0.6
requests==2.31.0
Pillow==10.0.1
python-dateutil==2.8.2
pytz==2023.3 